Separated more tests with files. Support raw shell command with !
